fix: UseParam fixes
