+++ src/app/m/location/page.tsx
@@ -1,22 +1,16 @@
-"use client"
-
-import {ScreenProvider} from "@/components/screen/screen-provider";
-import {ProtectedRoute} from "@/components/auth/ProtectedRoutes";
-import {LocationChangeContent} from "@/components/location-change/location-change-content";
-import {useSearchParams} from "next/navigation";
+import {Loader} from "@/components/ui/loader";
+import {Suspense} from "react";
+import LocationPageContent from "@/components/location-change/location-page";
 
 export default function LocationChangePage() {
-    const searchParams = useSearchParams();
-
-    const prev = searchParams.get("prev");
 
     return (
-        <ProtectedRoute>
-            <ScreenProvider>
-                <div className="w-full">
-                    <LocationChangeContent prev={prev!} />
-                </div>
-            </ScreenProvider>
-        </ProtectedRoute>
+        <Suspense fallback={
+            <div className="w-full h-screen flex items-center justify-center">
+                <Loader />
+            </div>
+        }>
+            <LocationPageContent />
+        </Suspense>
     )
 }
